--- pkg/netcertmanager/config/testdata/config-certmanager.yaml
@@ -1,1 +1,0 @@
-../../../../config/core/configmaps/certmanager.yaml
