+++ pkg/reconciler/certificate/config/testdata/config-certmanager.yaml
@@ -1,0 +1,1 @@
+../../../../../config/core/configmaps/certmanager.yaml
