+++ pkg/reconciler/certificate/config/zz_generated.deepcopy.go
