+++ pkg/reconciler/certificate/resources/cert_manager_certificate.go
@@ -18,6 +18,7 @@ package resources
 
 import (
 	"fmt"
+	"knative.dev/serving/pkg/reconciler/certificate/config"
 
 	cmv1 "github.com/cert-manager/cert-manager/pkg/apis/certmanager/v1"
 	cmeta "github.com/cert-manager/cert-manager/pkg/apis/meta/v1"
@@ -28,7 +29,6 @@ import (
 	netapi "knative.dev/networking/pkg/config"
 	"knative.dev/pkg/apis"
 	"knative.dev/pkg/kmeta"
-	"knative.dev/serving/pkg/netcertmanager/config"
 )
 
 const (
