+++ pkg/reconciler/certificate/resources/cert_manager_certificate_test.go
@@ -18,6 +18,7 @@ package resources
 
 import (
 	"fmt"
+	"knative.dev/serving/pkg/reconciler/certificate/config"
 	"strings"
 	"testing"
 
@@ -31,7 +32,6 @@ import (
 	"knative.dev/networking/pkg/apis/networking"
 	"knative.dev/networking/pkg/apis/networking/v1alpha1"
 	"knative.dev/pkg/kmeta"
-	"knative.dev/serving/pkg/netcertmanager/config"
 )
 
 const (
